Few changes in create groups

+++ remember_server/src/remember_service/GroupsService.java
@@ -12,7 +12,7 @@ public class GroupsService {
 	@Autowired
 	GroupsDao groupsDao;
 	
-	public int CreateGroup(GroupsDto group , Long id) {
+	public long CreateGroup(GroupsDto group , Long id) {
 		
 		return groupsDao.getGroupNameChecked(group , id);
 	}
